@@ -1,8 +1,7 @@
 ﻿using System.ComponentModel;
 
-namespace ASPNETCoreLocalization.Models
-{
-    public record Book(
-        [property: DisplayName("BookTitle")] string Title,
-        [property: DisplayName("Publisher")] string Publisher);
-}
+namespace ASPNETCoreLocalization.Models;
+
+public record Book(
+    [property: DisplayName("BookTitle")] string Title,
+    [property: DisplayName("Publisher")] string Publisher);
